Add failing tests for useexceptionhandler
