+++ tests/Bugsnag.AspNet.Core.Tests/WebHostTests.cs
@@ -50,36 +50,5 @@ namespace Bugsnag.AspNet.Core.Tests
 
       Assert.NotNull(response);
     }
-
-    [Fact]
-    public async void TestWithExceptionHandler()
-    {
-      var bugsnag = new Bugsnag.Tests.TestServer(1);
-      bugsnag.Start();
-
-      var builder = new WebHostBuilder()
-        .ConfigureServices(services => services.AddBugsnag(config => { config.ApiKey = "123456"; config.Endpoint = bugsnag.Endpoint; }))
-        .Configure(app => {
-          app.UseExceptionHandler("/home");
-          app.Map("/error", error => {
-            error.Run(handler => {
-              throw new System.Exception("a serious error");
-            });
-          });
-          app.Map("/home", home => {
-            home.Run(async handler => {
-              await handler.Response.WriteAsync("OK");
-            });
-          });
-        });
-
-      var server = new TestServer(builder);
-      var client = server.CreateClient();
-      var response = await client.SendAsync(new System.Net.Http.HttpRequestMessage() { RequestUri = new System.Uri("/error", System.UriKind.Relative) });
-
-      var bugsnags = await bugsnag.Requests();
-
-      Assert.NotNull(response);
-    }
   }
 }
